refactor to look more like offline conversions
