+++ packages/destination-actions/src/destinations/tiktok-conversions/utils.ts
@@ -1,0 +1,66 @@
+import { RequestClient } from '@segment/actions-core'
+import { Settings } from './generated-types'
+import { Payload } from './reportWebEvent/generated-types'
+import { formatEmails, formatPhones, formatUserIds } from './formatter'
+
+export function performWebEvent(request: RequestClient, settings: Settings, payload: Payload) {
+  const phone_numbers = formatPhones(payload.phone_number)
+  const emails = formatEmails(payload.email)
+  const userIds = formatUserIds(payload.external_id)
+
+  let payloadUrl, urlTtclid
+  if (payload.url) {
+    try {
+      payloadUrl = new URL(payload.url)
+    } catch (error) {
+      //  invalid url
+    }
+  }
+
+  if (payloadUrl) urlTtclid = payloadUrl.searchParams.get('ttclid')
+
+  return request('https://business-api.tiktok.com/open_api/v1.3/event/track/', {
+    method: 'post',
+    json: {
+      event_source: 'web',
+      event_source_id: settings.pixelCode,
+      partner_name: 'Segment',
+      data: [
+        {
+          event: payload.event,
+          event_time: payload.timestamp
+            ? Math.floor(new Date(payload.timestamp).getTime() / 1000)
+            : Math.floor(new Date().getTime() / 1000),
+          event_id: payload.event_id ? `${payload.event_id}` : undefined,
+          user: {
+            ttclid: payload.ttclid ? payload.ttclid : urlTtclid ? urlTtclid : undefined,
+            external_id: userIds,
+            phone: phone_numbers,
+            email: emails,
+            lead_id: payload.lead_id ? payload.lead_id : undefined,
+            ttp: payload.ttp ? payload.ttp : undefined,
+            ip: payload.ip ? payload.ip : undefined,
+            user_agent: payload.user_agent ? payload.user_agent : undefined,
+            locale: payload.locale ? payload.locale : undefined
+          },
+          properties: {
+            contents: payload.contents ? payload.contents : [],
+            content_type: payload.content_type ? payload.content_type : undefined,
+            currency: payload.currency ? payload.currency : undefined,
+            value: payload.value ? payload.value : undefined,
+            query: payload.query ? payload.query : undefined,
+            description: payload.description ? payload.description : undefined,
+            order_id: payload.order_id ? payload.order_id : undefined,
+            shop_id: payload.shop_id ? payload.shop_id : undefined
+          },
+          page: {
+            url: payload.url ? payload.url : undefined,
+            referrer: payload.referrer ? payload.referrer : undefined
+          },
+          limited_data_use: payload.limited_data_use ? payload.limited_data_use : false,
+          test_event_code: payload.test_event_code ? payload.test_event_code : undefined
+        }
+      ]
+    }
+  })
+}
